Install redux, begin search functionality
This commit installs Redux and begins connecting the Search bar to the store.

The API is not being hit yet. For some reason, the middleware makes the request
action but never hits either the success or failure. I suspect this is because
`redux-api-middleware@next` is not ready to be used on React Native.